+++ annotations/lib/realm_annotations.dart
@@ -3,8 +3,5 @@
 /// More dartdocs go here.
 library realm_annotations;
 
-export 'realm_property.dart';
+export 'src/realm_annotations_base.dart';
 export 'src/realm_types.dart';
-
-
-// TODO: Export any libraries intended for clients of this package.
