+++ annotations/lib/src/realm_annotations_base.dart
@@ -1,6 +1,38 @@
-// TODO: Put public facing types in this file.
+////////////////////////////////////////////////////////////////////////////////
+//
+// Copyright 2021 Realm Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+////////////////////////////////////////////////////////////////////////////////
 
-/// Checks if you are awesome. Spoiler: you are.
-class Awesome {
-  bool get isAwesome => true;
+class RealmModel {
+  const RealmModel();
+}
+
+class MapTo {
+  final String name;
+  const MapTo(this.name);
+}
+
+class PrimaryKey {
+  const PrimaryKey();
+}
+
+class Indexed {
+  const Indexed();
+}
+
+class Ignored {
+  const Ignored();
 }
